+++ termal_core/src/raw/mod.rs
@@ -1,8 +1,12 @@
+mod io_provider;
+mod stdio_provider;
 mod sys;
 mod terminal;
+mod wait_for_in;
 
-pub use sys::*;
-pub use terminal::*;
+pub use self::{
+    io_provider::*, stdio_provider::*, sys::*, terminal::*, wait_for_in::*,
+};
 
 #[cfg(feature = "events")]
 pub mod events;